[BS] add - Progres 31 July in Home

+++ Class/ChessPiece/Statistic.cs
@@ -3,14 +3,14 @@ namespace GameAutoChess.Class.ChessPiece;
 public class Statistic
 {
     public int HealthPiece { get; set; }
-    public int AttackPiece { get; set; }
+    public decimal AttackPiece { get; set; }
     public decimal AttackRate { get; set; }
     public int AttackRange { get; set; }
     public int DamageToEnemy { get; set; }
     public int Armor { get; set; }
     public int MagicResistance { get; set; } // Convert to % using /100
     
-    public Statistic(int healthPiece, int attackPiece, decimal attackRate, int attackRange, int damageToEnemy, int armor, int magicResistance)
+    public Statistic(int healthPiece, decimal attackPiece, decimal attackRate, int attackRange, int damageToEnemy, int armor, int magicResistance)
     {
         HealthPiece = healthPiece;
         AttackPiece = attackPiece;
